fix(wasm32-wasi): use custom resource directory
No longer relies on installing the compiler-rt sources into the clang default resource directory, which is often write protected (for good reason). Instead, the compiler-rt builtins for wasm32-wasi are loaded from the wasi-sdk cache directory. This allows for a more seamless getting started process.

Additionally, a check for the required version of clang is added. Version 17 or above should be able to compile for wasm32-wasi.